cbmc: Move contracts to separate header
CBMC annotations should be organized so that they clutter the
main code base as little as possible.

This commit moves the CBMC specs written so far from `poly.h`
into a new header `poly_specs.h` that's in the `cbmc/` subdirectory,
keeping the original `poly.h` from the reference implementation.

An alternative could be to make the annotations less cryptic
by introducing abbreviations like

```
    #define CONTRACT_ASSERT(...)      __CPROVER_assert(__VA_ARGS__)
    #define CONTRACT_ASSIGNS(...)     __CPROVER_assigns(__VA_ARGS__)
    #define CONTRACT_ASSIGNS_ERR(...) CONTRACT_ASSIGNS(__VA_ARGS__, _s2n_debug_info, s2n_errno)
    #define CONTRACT_ASSUME(...)      __CPROVER_assume(__VA_ARGS__)
    #define CONTRACT_REQUIRES(...)    __CPROVER_requires(__VA_ARGS__)
    #define CONTRACT_ENSURES(...)     __CPROVER_ensures(__VA_ARGS__)
    #define CONTRACT_INVARIANT(...)   __CPROVER_loop_invariant(__VA_ARGS__)
    #define CONTRACT_RETURN_VALUE     (__CPROVER_return_value)
    #define CONTRACT_ASSERT(...)
    #define CONTRACT_ASSIGNS(...)
    #define CONTRACT_ASSIGNS_ERR(...)
    #define CONTRACT_ASSUME(...)
    #define CONTRACT_REQUIRES(...)
    #define CONTRACT_ENSURES(...)
    #define CONTRACT_INVARIANT(...)
    #define CONTRACT_RETURN_VALUE
```

which is the approach taken by
s2n-tls (https://github.com/aws/s2n-tls/blob/ff610637368c31ef7bea5e3e5814c30af64263de/utils/s2n_ensure.h#L130).
