+++ cbmc/proofs/Makefile.common
@@ -319,6 +319,7 @@ ADD_LIBRARY_FLAG := --add-library
 # Preprocessor include paths -I...
 INCLUDES ?=
 INCLUDES += -I$(PROOFDIR)
+INCLUDES += -I$(PROOFDIR)/../specs
 INCLUDES += -I$(SRCDIR)/mlkem -I$(SRCDIR)/mlkem/sys
 INCLUDES += -I$(SRCDIR)/fips202
