+++ mlkem/poly.h
@@ -27,58 +27,24 @@ typedef struct
     int16_t coeffs[KYBER_N >> 1];
 } poly_mulcache;
 
-#define scalar_compress_q_16           KYBER_NAMESPACE(scalar_compress_q_16)
-#define scalar_decompress_q_16         KYBER_NAMESPACE(scalar_decompress_q_16)
-#define scalar_compress_q_32           KYBER_NAMESPACE(scalar_compress_q_32)
-#define scalar_decompress_q_32         KYBER_NAMESPACE(scalar_decompress_q_32)
+
+#define scalar_compress_q_16 KYBER_NAMESPACE(scalar_compress_q_16)
+uint32_t scalar_compress_q_16 (int32_t u);
+
+#define scalar_decompress_q_16 KYBER_NAMESPACE(scalar_decompress_q_16)
+uint32_t scalar_decompress_q_16 (uint32_t u);
+
+#define scalar_compress_q_32 KYBER_NAMESPACE(scalar_compress_q_32)
+uint32_t scalar_compress_q_32 (int32_t u);
+
+#define scalar_decompress_q_32 KYBER_NAMESPACE(scalar_decompress_q_32)
+uint32_t scalar_decompress_q_32 (uint32_t u);
+
 #define scalar_signed_to_unsigned_q_16 KYBER_NAMESPACE(scalar_signed_to_unsigned_q_16)
-
-uint32_t scalar_compress_q_16   (int32_t u)
-/* INDENT-OFF */
-__CPROVER_requires(0 <= u && u < KYBER_Q)
-__CPROVER_ensures(__CPROVER_return_value < 16)
-__CPROVER_ensures(__CPROVER_return_value == (((uint32_t) u * 16 + KYBER_Q / 2) / KYBER_Q) % 16);
-/* INDENT-ON */
-
-uint32_t scalar_decompress_q_16 (uint32_t u)
-/* INDENT-OFF */
-__CPROVER_requires(0 <= u && u < 16)
-__CPROVER_ensures(__CPROVER_return_value < KYBER_Q);
-/* INDENT-ON */
-
-uint32_t scalar_compress_q_32   (int32_t u)
-/* INDENT-OFF */
-__CPROVER_requires(0 <= u && u < KYBER_Q)
-__CPROVER_ensures(__CPROVER_return_value < 32)
-__CPROVER_ensures(__CPROVER_return_value == (((uint32_t) u * 32 + KYBER_Q / 2) / KYBER_Q) % 32);
-/* INDENT-ON */
-
-uint32_t scalar_decompress_q_32 (uint32_t u)
-/* INDENT-OFF */
-__CPROVER_requires(0 <= u && u < 32)
-__CPROVER_ensures(__CPROVER_return_value < KYBER_Q);
-/* INDENT-ON */
-
-uint16_t scalar_signed_to_unsigned_q_16 (int16_t c)
-/* *INDENT-OFF* */
-__CPROVER_requires(c > -KYBER_Q) // c >= -3328
-__CPROVER_requires(c < KYBER_Q)  // c <= 3328
-__CPROVER_ensures(__CPROVER_return_value >= 0)
-__CPROVER_ensures(__CPROVER_return_value < KYBER_Q)
-__CPROVER_ensures(__CPROVER_return_value == (int32_t) c + (((int32_t) c < 0) * KYBER_Q));
-/* *INDENT-ON* */
+uint16_t scalar_signed_to_unsigned_q_16 (int16_t c);
 
 #define poly_compress KYBER_NAMESPACE(poly_compress)
-void poly_compress(uint8_t r[KYBER_POLYCOMPRESSEDBYTES], const poly *a)
-/* *INDENT-OFF* */
-__CPROVER_requires(r != NULL)
-__CPROVER_requires(__CPROVER_is_fresh(r, KYBER_POLYCOMPRESSEDBYTES))
-__CPROVER_requires(a != NULL)
-__CPROVER_requires(__CPROVER_is_fresh(a, sizeof(poly)))
-__CPROVER_requires(__CPROVER_forall { unsigned k; (k < KYBER_N) ==> ( -KYBER_Q < a->coeffs[k] && a->coeffs[k] < KYBER_Q ) })
-__CPROVER_assigns(__CPROVER_object_whole(r));
-/* *INDENT-ON* */
-
+void poly_compress(uint8_t r[KYBER_POLYCOMPRESSEDBYTES], const poly *a);
 
 #define poly_decompress KYBER_NAMESPACE(poly_decompress)
 void poly_decompress(poly *r, const uint8_t a[KYBER_POLYCOMPRESSEDBYTES]);
